apps/api-python/pyproject.toml
[project]
name = "api-python"
version = "0.1.0"
description = "Seed OI API"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.2",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "structlog>=24.1.0",
    "asgi-correlation-id>=4.3.1",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
]
